+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- Package now supports `2.0` of the `laravel-json-api/core` dependency.
+
 ## [1.0.0] - 2021-07-31
 
 Initial stable release, with no changes from `1.0.0-beta.2`.