+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 
 - Package now supports Laravel 9.
 - Package now supports `2.0` of the `laravel-json-api/core` dependency.
+- [#5](https://github.com/laravel-json-api/spec/pull/5) Add Italian translation files.
 
 ## [1.0.0] - 2021-07-31